Make automatic respawn (from lap cheat prevention) more lenient
- When potential lap cheating is detected, start a 10
  second timer (previously 1 second)
- After 3 seconds, tell the player they can respawn with
  the Y button
  - You don't need to hold the Y button, just tap it
- After 10 seconds, automatically respawns the player

+++ src/d_player.h
@@ -671,7 +671,7 @@ struct player_t
 	mobj_t *ringShooter;	// DEZ respawner object
 	tic_t airtime; 			// Used to track just air time, but has evolved over time into a general "karted" timer. Rename this variable?
 	tic_t lastairtime;
-	UINT8 bigwaypointgap;	// timer counts down if finish line distance gap is too big to update waypoint
+	UINT16 bigwaypointgap;	// timer counts down if finish line distance gap is too big to update waypoint
 	UINT8 startboost;		// (0 to 125) - Boost you get from start of race
 	UINT8 dropdashboost;	// Boost you get when holding A while respawning
